v2.0.0, refactored EventStoreService so it no longer connects to the EventStore on its own but rather relies on dependency injection to get an IEventStore-instance, added some documentation
