+++ HiP-EventStoreLib/EventStoreLlp/EventStore.cs
@@ -2,6 +2,9 @@
 using EventStore.ClientAPI;
 using System.Threading.Tasks;
 using Nito.AsyncEx;
+using Microsoft.Extensions.Options;
+using System;
+using System.Diagnostics;
 
 namespace PaderbornUniversity.SILab.Hip.EventSourcing.EventStoreLlp
 {
@@ -21,6 +24,30 @@ namespace PaderbornUniversity.SILab.Hip.EventSourcing.EventStoreLlp
         public EventStore(IEventStoreConnection underlyingConnection)
         {
             UnderlyingConnection = underlyingConnection;
+        }
+
+        public EventStore(IOptions<EventStoreConfig> config)
+        {
+            // Validate config
+            if (string.IsNullOrWhiteSpace(config.Value.Host))
+                throw new ArgumentException($"Invalid configuration: Missing value for '{nameof(config.Value.Host)}'");
+
+            // Prevent accidentally working with a production database
+            if (Debugger.IsAttached)
+            {
+                Debug.Assert(config.Value.Host.Contains("localhost"),
+                    "It looks like you are trying to connect to a production Event Store database. Are you sure you wish to continue?");
+            }
+
+            var settings = ConnectionSettings.Create()
+                .EnableVerboseLogging()
+                .Build();
+
+            var uri = new Uri(config.Value.Host);
+            var connection = EventStoreConnection.Create(settings, uri);
+            connection.ConnectAsync().Wait();
+
+            UnderlyingConnection = connection;
         }
 
         internal async Task DeleteStreamAsync(string name)
